+++ dbms/src/Storages/DeltaMerge/CMakeLists.txt
@@ -29,6 +29,7 @@ add_headers_and_sources(delta_merge ./Delta)
 add_headers_and_sources(delta_merge ./ReadThread)
 add_headers_and_sources(delta_merge ./Remote)
 add_headers_and_sources(delta_merge ./Remote/DataStore)
+add_headers_and_sources(delta_merge ./Decode)
 
 add_library(delta_merge ${delta_merge_headers} ${delta_merge_sources})
 target_link_libraries(delta_merge PRIVATE dbms page)
