+++ dbms/src/Storages/KVStore/MultiRaft/IngestSST.cpp
@@ -16,7 +16,7 @@
 #include <Common/TiFlashMetrics.h>
 #include <Common/setThreadName.h>
 #include <Interpreters/Context.h>
-#include <Storages/DeltaMerge/SSTFilesToDTFilesOutputStream.h>
+#include <Storages/DeltaMerge/Decode/SSTFilesToDTFilesOutputStream.h>
 #include <Storages/KVStore/Decode/RegionTable.h>
 #include <Storages/KVStore/FFI/ProxyFFI.h>
 #include <Storages/KVStore/FFI/SSTReader.h>
